Massive update which breaks the git model
For the past 12 months, I've completely disregarded git as a
source management system for vivi.  Instead, I merely used it to
"save" my work (checking out out the previous version of a file if
the current work wasn't working) and to synchronize it between
computers.  Most commit messages were "asdf".  That was not a good
thing, but I just don't have the energy to go through all 330
commits from the past 12 months to reorganize them into logical
commits.  Hence the massive update.

I forsee using good source management in this project in the
future.